UI  - life 구현.  - life는 현재 최대체력만큼의 빈하트를 가지고 단위가 1하트.  - 원작을 따라가려면 반하트 단위로 변경이 필요.
PlayerTable.csv
 - Player의 타입에 따라서 스탯의 변경이 가능하게 미리 테이블 생성.

SceneTitle
 - the binding of isaac 로고 구현.

+++ My_Isaac/GameObjects/Tear.cpp
@@ -58,7 +58,7 @@ void Tear::Update(float dt)
 	}
 }
 
-void Tear::Shoot(const sf::Vector2f& position, const sf::Vector2f& direction, float speed, float damage)
+void Tear::Shoot(const sf::Vector2f& position, const sf::Vector2f& direction, float speed, int damage)
 {
 	SetPosition(position);
